+++ src/Pim/Bundle/CatalogBundle/Builder/ProductTemplateBuilder.php
@@ -1,28 +1,21 @@
 <?php
 
-namespace Pim\Bundle\CatalogBundle\Manager;
+namespace Pim\Bundle\CatalogBundle\Builder;
 
-use Pim\Bundle\CatalogBundle\Builder\ProductBuilder;
 use Pim\Bundle\CatalogBundle\Model\AttributeInterface;
-use Pim\Bundle\CatalogBundle\Model\GroupInterface;
 use Pim\Bundle\CatalogBundle\Model\ProductTemplateInterface;
 use Pim\Bundle\CatalogBundle\Model\ProductValueInterface;
-use Pim\Bundle\CatalogBundle\Repository\AttributeRepositoryInterface;
 use Symfony\Component\Serializer\Normalizer\DenormalizerInterface;
 use Symfony\Component\Serializer\Normalizer\NormalizerInterface;
 
 /**
- * Product template attributes manager
+ * Product template builder, allows to create new product template and update them
  *
  * @author    Filips Alpe <filips@akeneo.com>
  * @copyright 2015 Akeneo SAS (http://www.akeneo.com)
  * @license   http://opensource.org/licenses/osl-3.0.php  Open Software License (OSL 3.0)
- *
- * TODO not comfortable with the naming of ProductTemplateAttributesManager we could maybe merge
- * ProductTemplateAttributesManager and ProductTemplateFactory to a ProductTemplateBuilder
- * with create(), addAttributes(), removeAttribute() methods
  */
-class ProductTemplateAttributesManager
+class ProductTemplateBuilder implements ProductTemplateBuilderInterface
 {
     /** @var NormalizerInterface */
     protected $normalizer;
@@ -33,66 +26,43 @@ class ProductTemplateAttributesManager
     /** @var ProductBuilder */
     protected $productBuilder;
 
-    /** @var AttributeRepositoryInterface */
-    protected $attributeRepository;
+    /** @var string */
+    protected $productTemplateClass;
 
     /** @var string */
     protected $productClass;
 
     /**
-     * @param NormalizerInterface          $normalizer
-     * @param DenormalizerInterface        $denormalizer
-     * @param ProductBuilder               $productBuilder
-     * @param AttributeRepositoryInterface $attributeRepository
-     * @param string                       $productClass
+     * @param NormalizerInterface   $normalizer
+     * @param DenormalizerInterface $denormalizer
+     * @param ProductBuilder        $productBuilder
+     * @param string                $productTemplateClass
+     * @param string                $productClass
      */
     public function __construct(
         NormalizerInterface $normalizer,
         DenormalizerInterface $denormalizer,
         ProductBuilder $productBuilder,
-        AttributeRepositoryInterface $attributeRepository,
+        $productTemplateClass,
         $productClass
     ) {
-        $this->normalizer          = $normalizer;
-        $this->denormalizer        = $denormalizer;
-        $this->productBuilder      = $productBuilder;
-        $this->attributeRepository = $attributeRepository;
-        $this->productClass        = $productClass;
+        $this->normalizer           = $normalizer;
+        $this->denormalizer         = $denormalizer;
+        $this->productBuilder       = $productBuilder;
+        $this->productTemplateClass = $productTemplateClass;
+        $this->productClass         = $productClass;
     }
 
     /**
-     * Get non eligible attributes to a product template
-     *
-     * @param GroupInterface $group
-     *
-     * @return AttributeInterface[]
+     * {@inheritdoc}
      */
-    public function getNonEligibleAttributes(GroupInterface $group)
+    public function createProductTemplate()
     {
-        $attributes = $group->getAxisAttributes()->toArray();;
-
-        $template = $group->getProductTemplate();
-        if (null !== $template) {
-            foreach (array_keys($template->getValuesData()) as $attributeCode) {
-                $attributes[] = $this->attributeRepository->findOneByIdentifier($attributeCode);
-            }
-        }
-
-        $uniqueAttributes = $this->attributeRepository->findBy(['unique' => true]);
-        foreach ($uniqueAttributes as $attribute) {
-            if (!in_array($attribute, $attributes)) {
-                $attributes[] = $attribute;
-            }
-        }
-
-        return $attributes;
+        return new $this->productTemplateClass();
     }
 
     /**
-     * Add required value(s) that link an attribute to a product template
-     *
-     * @param ProductTemplateInterface $template
-     * @param AttributeInterface[]     $attributes
+     * {@inheritdoc}
      */
     public function addAttributes(ProductTemplateInterface $template, array $attributes)
     {
@@ -102,10 +72,7 @@ class ProductTemplateAttributesManager
     }
 
     /**
-     * Delete values that link an attribute to the product template
-     *
-     * @param ProductTemplateInterface $template
-     * @param AttributeInterface       $attribute
+     * {@inheritdoc}
      */
     public function removeAttribute(ProductTemplateInterface $template, AttributeInterface $attribute)
     {