+++ src/Pim/Bundle/CatalogBundle/Query/ProductQueryBuilder.php
@@ -105,7 +105,6 @@ class ProductQueryBuilder implements ProductQueryBuilderInterface
     {
         $attribute = $this->attributeRepository->findOneBy(['code' => FieldFilterHelper::getCode($field)]);
 
-        //TODO: make this transparent by adding it the filter registry
         if ($attribute !== null) {
             $filter = $this->filterRegistry->getAttributeFilter($attribute);
         } else {
@@ -143,7 +142,6 @@ class ProductQueryBuilder implements ProductQueryBuilderInterface
     {
         $attribute = $this->attributeRepository->findOneBy(['code' => $field]);
 
-        //TODO: make this transparent by adding it the sorter registry
         if ($attribute !== null) {
             $sorter = $this->sorterRegistry->getAttributeSorter($attribute);
         } else {