+++ src/Pim/Bundle/EnrichBundle/Controller/FamilyController.php
@@ -299,7 +299,7 @@ class FamilyController extends AbstractDoctrineController
         return $this->createForm(
             'pim_available_attributes',
             $availableAttributes ?: new AvailableAttributes(),
-            array('attributes' => $attributes)
+            array('excluded_attributes' => $attributes)
         );
     }
 }