+++ src/Pim/Bundle/EnrichBundle/Controller/VariantGroupController.php
@@ -3,10 +3,12 @@
 namespace Pim\Bundle\EnrichBundle\Controller;
 
 use Oro\Bundle\SecurityBundle\Annotation\AclAncestor;
+use Pim\Bundle\CatalogBundle\Builder\ProductTemplateBuilderInterface;
 use Pim\Bundle\CatalogBundle\Entity\Group;
 use Pim\Bundle\CatalogBundle\Factory\GroupFactory;
 use Pim\Bundle\CatalogBundle\Manager\GroupManager;
-use Pim\Bundle\CatalogBundle\Manager\ProductTemplateAttributesManager;
+use Pim\Bundle\CatalogBundle\Manager\ProductTemplateBuilder;
+use Pim\Bundle\CatalogBundle\Manager\VariantGroupAttributesResolver;
 use Pim\Bundle\CatalogBundle\Model\AvailableAttributes;
 use Pim\Bundle\CatalogBundle\Model\GroupInterface;
 use Pim\Bundle\CatalogBundle\Repository\AttributeRepositoryInterface;
@@ -36,24 +38,28 @@ class VariantGroupController extends GroupController
     /** @var AttributeRepositoryInterface */
     protected $attributeRepository;
 
-    /** @var ProductTemplateAttributesManager */
-    protected $tplAttributesManager;
+    /** @var ProductTemplateBuilder */
+    protected $templateBuilder;
+
+    /** @var VariantGroupAttributesResolver */
+    protected $groupAttributesResolver;
 
     /**
-     * @param Request                          $request
-     * @param EngineInterface                  $templating
-     * @param RouterInterface                  $router
-     * @param SecurityContextInterface         $securityContext
-     * @param FormFactoryInterface             $formFactory
-     * @param ValidatorInterface               $validator
-     * @param TranslatorInterface              $translator
-     * @param EventDispatcherInterface         $eventDispatcher
-     * @param GroupManager                     $groupManager
-     * @param HandlerInterface                 $groupHandler
-     * @param Form                             $groupForm
-     * @param GroupFactory                     $groupFactory
-     * @param AttributeRepositoryInterface     $attributeRepository
-     * @param ProductTemplateAttributesManager $tplAttributesManager
+     * @param Request                        $request
+     * @param EngineInterface                $templating
+     * @param RouterInterface                $router
+     * @param SecurityContextInterface       $securityContext
+     * @param FormFactoryInterface           $formFactory
+     * @param ValidatorInterface             $validator
+     * @param TranslatorInterface            $translator
+     * @param EventDispatcherInterface       $eventDispatcher
+     * @param GroupManager                   $groupManager
+     * @param HandlerInterface               $groupHandler
+     * @param Form                           $groupForm
+     * @param GroupFactory                   $groupFactory
+     * @param AttributeRepositoryInterface   $attributeRepository
+     * @param ProductTemplateBuilder         $templateBuilder
+     * @param VariantGroupAttributesResolver $groupAttributesResolver
      */
     public function __construct(
         Request $request,
@@ -69,7 +75,8 @@ class VariantGroupController extends GroupController
         Form $groupForm,
         GroupFactory $groupFactory,
         AttributeRepositoryInterface $attributeRepository,
-        ProductTemplateAttributesManager $tplAttributesManager
+        ProductTemplateBuilderInterface $templateBuilder,
+        VariantGroupAttributesResolver $groupAttributesResolver
     ) {
         parent::__construct(
             $request,
@@ -87,7 +94,8 @@ class VariantGroupController extends GroupController
         );
 
         $this->attributeRepository = $attributeRepository;
-        $this->tplAttributesManager = $tplAttributesManager;
+        $this->templateBuilder = $templateBuilder;
+        $this->groupAttributesResolver = $groupAttributesResolver;
     }
 
     /**
@@ -174,8 +182,7 @@ class VariantGroupController extends GroupController
         return $this->createForm(
             'pim_available_attributes',
             new AvailableAttributes(),
-            // TODO: this key is really not well named...
-            ['attributes' => $this->tplAttributesManager->getNonEligibleAttributes($group)]
+            ['excluded_attributes' => $this->groupAttributesResolver->getNonEligibleAttributes($group)]
         );
     }
 }